move types/geometry to geometry crate
