+++ Cargo.toml
@@ -11,6 +11,7 @@ members = [
   "crates/control",
   "crates/filtering",
   "crates/framework",
+  "crates/geometry",
   "crates/hardware",
   "crates/hulk",
   "crates/hulk_nao",
@@ -82,6 +83,7 @@ filtering = { path = "crates/filtering" }
 framework = { path = "crates/framework" }
 futures-util = "0.3.24"
 fuzzy-matcher = "0.3.7"
+geometry = { path = "crates/geometry" }
 gilrs = "0.10.1"
 glob = "0.3.0"
 hardware = { path = "crates/hardware" }
