+++ crates/control/Cargo.toml
@@ -12,16 +12,17 @@ color-eyre = { workspace = true }
 context_attribute = { workspace = true }
 filtering = { workspace = true }
 framework = { workspace = true }
+geometry = { workspace = true }
 hardware = { workspace = true }
 itertools = { workspace = true }
 kinematics = { workspace = true }
 log = { workspace = true }
 motionfile = { workspace = true }
 nalgebra = { workspace = true }
+num-traits = {workspace = true}
 ordered-float = { workspace = true }
 projection = { workspace = true }
 rand = {workspace = true}
-num-traits = {workspace = true}
 serde = { workspace = true }
 serialize_hierarchy = { workspace = true }
 smallvec = { workspace = true }
