+++ crates/geometry/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "geometry"
+version = "0.1.0"
+edition = "2021"
+license = "GPL-3.0-only"
+homepage = "https://github.com/hulks/hulk"
+
+[dependencies]
+approx = { workspace = true }
+nalgebra = { workspace = true }
+serde = { workspace = true }
+serialize_hierarchy = { workspace = true }
