+++ crates/geometry/src/lib.rs
@@ -1,0 +1,8 @@
+pub mod arc;
+pub mod circle;
+pub mod circle_tangents;
+pub mod line_segment;
+pub mod look_at;
+pub mod orientation;
+pub mod rectangle;
+pub mod two_line_segments;
