+++ crates/types/Cargo.toml
@@ -9,6 +9,7 @@ homepage = "https://github.com/hulks/hulk"
 approx = { workspace = true }
 color-eyre = { workspace = true }
 enum-iterator = { workspace = true }
+geometry = { workspace = true }
 image = { workspace = true }
 nalgebra = { workspace = true }
 ordered-float = { workspace = true }
