+++ crates/types/src/ball.rs
@@ -1,9 +1,8 @@
 use approx::{AbsDiffEq, RelativeEq};
+use geometry::circle::Circle;
 use nalgebra::Point2;
 use serde::{Deserialize, Serialize};
 use serialize_hierarchy::SerializeHierarchy;
-
-use super::geometry::Circle;
 
 #[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize, SerializeHierarchy)]
 pub struct CandidateEvaluation {
