+++ crates/types/src/planned_path.rs
@@ -1,9 +1,8 @@
 use approx::{AbsDiffEq, RelativeEq};
+use geometry::{arc::Arc, line_segment::LineSegment, orientation::Orientation};
 use nalgebra::{Isometry2, Point2};
 use serde::{Deserialize, Serialize};
 use serialize_hierarchy::SerializeHierarchy;
-
-use super::geometry::{Arc, LineSegment, Orientation};
 
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq, SerializeHierarchy)]
 pub enum PathSegment {
