+++ crates/types/src/rule_obstacles.rs
@@ -1,7 +1,7 @@
+use geometry::{circle::Circle, rectangle::Rectangle};
 use serde::{Deserialize, Serialize};
 use serialize_hierarchy::SerializeHierarchy;
 
-use crate::geometry::{Circle, Rectangle};
 #[derive(Clone, Copy, Debug, Deserialize, Serialize, SerializeHierarchy)]
 pub enum RuleObstacle {
     Circle(Circle),
