+++ crates/vision/Cargo.toml
@@ -13,6 +13,7 @@ context_attribute = { workspace = true }
 fast_image_resize = { workspace = true }
 filtering = { workspace = true }
 framework = { workspace = true }
+geometry = { workspace = true }
 hardware = { workspace = true }
 itertools = { workspace = true }
 nalgebra = { workspace = true }
