+++ tools/twix/src/twix_painter.rs
@@ -5,12 +5,9 @@ use eframe::{
     emath::{Pos2, Rect},
     epaint::{Color32, PathShape, Rounding, Shape, Stroke},
 };
+use geometry::{arc::Arc, circle::Circle, orientation::Orientation};
 use nalgebra::{point, vector, Isometry2, Point2, Rotation2, SMatrix, Similarity2, Vector2};
-use types::{
-    field_dimensions::FieldDimensions,
-    geometry::{Arc, Circle, Orientation},
-    planned_path::PathSegment,
-};
+use types::{field_dimensions::FieldDimensions, planned_path::PathSegment};
 
 pub enum CoordinateSystem {
     RightHand,
